backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
PyMuPDF==1.23.8
sentence-transformers==2.2.2
chromadb==0.4.22
numpy==1.24.3